chore: move coverageExclusions
